rm pose, move transforms
